Remove new projects. Merge changes to existing project
