--- Altinn.Platform.Events.IsolatedFunctions/Properties/launchSettings.json
@@ -1,12 +1,0 @@
-{
-  "profiles": {
-    "Altinn.Platform.Events.FunctionsOutbound": {
-      "commandName": "Project",
-      "launchBrowser": false,
-      "environmentVariables": {
-        "AZURE_FUNCTIONS_ENVIRONMENT": "Development",
-        "DOTNET_ENVIRONMENT": "Development"
-      }
-    }
-  }
-}
