--- Altinn.Platform.Events.IsolatedFunctions/Properties/serviceDependencies.json
@@ -1,19 +1,0 @@
-{
-  "dependencies": {
-    "appInsights1": {
-      "type": "appInsights"
-    },
-    "storage1": {
-      "type": "storage",
-      "connectionId": "AzureWebJobsStorage"
-    },
-    "secrets1": {
-      "type": "secrets"
-    },
-    "storage2": {
-      "type": "storage",
-      "connectionId": "StorageConnection",
-      "dynamicId": null
-    }
-  }
-}
