--- Altinn.Platform.Events.IsolatedFunctions/Properties/serviceDependencies.local.json
@@ -1,20 +1,0 @@
-{
-  "dependencies": {
-    "appInsights1": {
-      "type": "appInsights.sdk"
-    },
-    "storage1": {
-      "type": "storage.emulator",
-      "connectionId": "AzureWebJobsStorage"
-    },
-    "secrets1": {
-      "type": "secrets.user"
-    },
-    "storage2": {
-      "secretStore": "LocalSecretsFile",
-      "type": "storage.emulator",
-      "connectionId": "StorageConnection",
-      "dynamicId": null
-    }
-  }
-}
