--- Altinn.Platform.Events.IsolatedFunctions/local.settings.json
@@ -1,10 +1,0 @@
-{
-  "IsEncrypted": false,
-  "Values": {
-    "AzureWebJobsStorage": "UseDevelopmentStorage=true",
-    "FUNCTIONS_WORKER_RUNTIME": "dotnet-isolated",
-    "QueueStorage": "DefaultEndpointsProtocol=http;AccountName=devstoreaccount1;AccountKey=Eby8vdM02xNOcqFlqUwJPLlmEtlCDXJ1OUzFT50uSRZ6IFsuFq2UVErCz4I6tq/K1SZFPTOtr/KBHBeksoGMGw==;QueueEndpoint=http://127.0.0.1:10001/devstoreaccount1",
-    "Platform:ApiEventsEndpoint": "http://localhost:5080/events/api/v1/",
-    "KeyVault:KeyVaultURI": "https://altinn-at21-kv.vault.azure.net/"
-  }
-}
