+++ Altinn.Platform.Events.sln
@@ -13,10 +13,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Altinn.Platform.Events.Func
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DbTools", "src\DbTools\DbTools.csproj", "{926F89EE-00F2-49F6-991D-32ABD3610C05}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Altinn.Platform.Events.IsolatedFunctions", "Altinn.Platform.Events.IsolatedFunctions\Altinn.Platform.Events.IsolatedFunctions.csproj", "{8772FA38-D31A-420C-BF14-61803D2C152E}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Altinn.Platform.Events.IsolatedFunctions.Tests", "Altinn.Platform.Events.IsolatedFunctions.Tests\Altinn.Platform.Events.IsolatedFunctions.Tests.csproj", "{B68B1CE7-FD9E-46FE-B13D-E7CB9C4AB613}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -43,14 +39,6 @@ Global
 		{926F89EE-00F2-49F6-991D-32ABD3610C05}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{926F89EE-00F2-49F6-991D-32ABD3610C05}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{926F89EE-00F2-49F6-991D-32ABD3610C05}.Release|Any CPU.Build.0 = Release|Any CPU
-		{8772FA38-D31A-420C-BF14-61803D2C152E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{8772FA38-D31A-420C-BF14-61803D2C152E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8772FA38-D31A-420C-BF14-61803D2C152E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{8772FA38-D31A-420C-BF14-61803D2C152E}.Release|Any CPU.Build.0 = Release|Any CPU
-		{B68B1CE7-FD9E-46FE-B13D-E7CB9C4AB613}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B68B1CE7-FD9E-46FE-B13D-E7CB9C4AB613}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B68B1CE7-FD9E-46FE-B13D-E7CB9C4AB613}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B68B1CE7-FD9E-46FE-B13D-E7CB9C4AB613}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
