+++ src/Events.Functions/Altinn.Platform.Events.Functions.csproj
@@ -1,17 +1,28 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <TargetFramework>net9.0</TargetFramework>
+    <AzureFunctionsVersion>v4</AzureFunctionsVersion>
     <GenerateDocumentationFile>true</GenerateDocumentationFile>
     <ProjectGuid>{1F3A1E44-1333-46B8-8A70-08BE01C074F6}</ProjectGuid>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="Altinn.Common.AccessTokenClient" Version="1.1.5" />
-    <PackageReference Include="Azure.Identity" Version="1.14.1" />
-    <PackageReference Include="Azure.Security.KeyVault.Secrets" Version="4.8.0" />
-    <PackageReference Include="Azure.Security.KeyVault.Certificates" Version="4.8.0" />
-    <PackageReference Include="CloudNative.CloudEvents.SystemTextJson" Version="2.8.0" />
-    <PackageReference Include="Microsoft.Extensions.Http" Version="9.0.8" />
-    <PackageReference Include="Microsoft.ApplicationInsights" Version="2.23.0" />
+	  <PackageReference Include="Azure.Data.Tables" Version="12.11.0" />
+	  <PackageReference Include="Azure.Storage.Blobs" Version="12.25.0" />
+	  <PackageReference Include="Azure.Storage.Files.Shares" Version="12.23.0" />
+	  <PackageReference Include="Azure.Storage.Queues" Version="12.23.0" />
+	  <PackageReference Include="Microsoft.Azure.Functions.Worker" Version="2.0.0" />
+	  <PackageReference Include="Microsoft.Azure.Functions.Worker.Extensions.Http.AspNetCore" Version="2.0.2" />
+	  <PackageReference Include="Microsoft.Azure.Functions.Worker.Extensions.Storage.Queues" Version="5.5.2" />
+	  <PackageReference Include="Microsoft.Azure.Functions.Worker.Sdk" Version="2.0.5" />
+	  <PackageReference Include="Microsoft.Extensions.Azure" Version="1.12.0" />
+	  <PackageReference Include="Microsoft.Extensions.Configuration.UserSecrets" Version="9.0.8" />
+
+	  <PackageReference Include="Altinn.Common.AccessTokenClient" Version="3.0.15" />
+	  <PackageReference Include="Azure.Security.KeyVault.Secrets" Version="4.8.0" />
+	  <PackageReference Include="Azure.Security.KeyVault.Certificates" Version="4.8.0" />
+	  <PackageReference Include="CloudNative.CloudEvents.SystemTextJson" Version="2.8.0" />
+	  <PackageReference Include="Microsoft.Extensions.Http" Version="9.0.8" />
+	  <PackageReference Include="Microsoft.ApplicationInsights" Version="2.23.0" />
   </ItemGroup>
 
   <ItemGroup Condition="'$(Configuration)'=='Debug'">
@@ -19,9 +30,23 @@
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
     </PackageReference>
-    <AdditionalFiles Include="..\..\stylecop.json">
-      <Link>stylecop.json</Link>
-    </AdditionalFiles>
+	<AdditionalFiles Include="..\..\stylecop.json">
+		<Link>stylecop.json</Link>
+	</AdditionalFiles>
+  </ItemGroup>
+
+  <PropertyGroup>
+    <_FunctionsSkipCleanOutput>true</_FunctionsSkipCleanOutput>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <None Update="host.json">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </None>
+    <None Update="local.settings.json">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+      <CopyToPublishDirectory>Never</CopyToPublishDirectory>
+    </None>
   </ItemGroup>
 
   <ItemGroup>
