+++ test/Altinn.Platform.Events.Functions.Tests/IntegrationTests/RequiresAzuriteFactAttribute.cs
@@ -1,0 +1,20 @@
+using Xunit;
+
+namespace Altinn.Platform.Events.Functions.Tests.IntegrationTests;
+
+[AttributeUsage(AttributeTargets.Method, AllowMultiple = false)]
+public sealed class RequiresAzuriteFactAttribute : FactAttribute
+{
+    public RequiresAzuriteFactAttribute()
+    {
+        if (!AzuriteTestsEnabled())
+        {
+            Skip = "Skipped: These tests require Azurite running on a local machine).";
+        }
+    }
+
+    private static bool AzuriteTestsEnabled()
+    {
+        return false;
+    }
+}
